requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.2.3
django-otp==1.2.2
djangorestframework==3.14.0
drf-yasg==1.21.6
environ==1.0
executing==1.2.0
inflection==0.5.1
ipython==8.12.2
jedi==0.18.2
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyotp==2.8.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==4.7.1
uritemplate==4.1.1
wcwidth==0.2.6
